FTP
the project about a client server that work on FTP protocol that we implemented.

for more information on the protocol and how does it work,
you can read the background in the "network project.doc".

for instruction on how to use the client server project to demonstrate the protocol,
there is specific explanation in the user guide in the "network project.doc".

all the diagrams build in UMLet, and found in "diagrams" folder.

Ftp stands for file transfer protocol, the ftp is a standard network communication that used for transferring files across the networks over the tcp/ip, it's based on the same concept of the server-client architecture, in order to transfer a file, you need the sender that is a client and a server that receive the file.  The implementation of the ftp protocol dates from 1971, described in RFC 141, when a file transfer system between MIT machines was developed. Since the 1971 many improvements has been done to the original ftp protocol, in this days the current ftp protocol  defined by the RFC 959. The ftp protocol become one of our important protocol that we use it in our daily tasks without even knowing that, it may be in file sharing between remote machines or independence between client and server machine system file.